+++ components/notifier/src/models/notification.py
@@ -1,7 +1,5 @@
 """Notification."""
 
-from datetime import datetime, timedelta
-
 from models.metric_notification_data import MetricNotificationData
 
 
@@ -15,22 +13,14 @@ class Notification:
         self.metrics: list[MetricNotificationData] = metrics
         self.destination_uuid = destination_uuid
         self.destination = destination
-        self.creation_time = datetime.now()
 
     def __eq__(self, other):
-        """Check if the notification itself is the same, regardless of it's metric content."""
+        """Check if the notification itself is the same, regardless of its metric content."""
         return (
             self.report_title == other.report_title
             and self.destination_uuid == other.destination_uuid
             and self.destination == other.destination
         )
-
-    def ready(self):
-        """Return whether this notification can be sent."""
-        age = datetime.now() - self.creation_time
-        notification_frequency = self.destination.get("frequency", 0)
-        minimal_age = timedelta(minutes=int(notification_frequency))
-        return age >= minimal_age
 
     def merge_notification(self, new_metrics):
         """Merge new metrics into this notification."""